♻️ refactor(server): move server/ to internal/server/
Relocate the HTTP server package under internal/ to match the project's
package layout convention. Update import paths and fix relative test paths.
