+++ internal/server/CLAUDE.md
