+++ internal/server/manifest_plugins.go
