+++ internal/server/plugins.go
