+++ internal/server/scheduler_helpers_test.go
