+++ internal/server/sessions_test.go
