+++ internal/server/skills_scoped.go
