+++ internal/server/skills_upload.go
