+++ internal/server/vault_test.go
