+++ internal/server/weixin_qr_test.go
